Cleaned up fallaback reverse geocode observable.
